+++ GameTemplate/Game/EnemyStateSideSlash.cpp
@@ -5,25 +5,16 @@
 #include "Player.h"
 
 #include "EnemyParameter.h"
+#include "PlayerParameter.h"
 
-void EnemyStateSideSlash::Start(Enemy* enemy)
+void EnemyStateSideSlash::Start(Enemy* enemy, Player* player)
 {
 	m_isPlayAnimation = true;
 
 	enemy->GetModel()->PlayAnimation(Enemy::enAnimationClip_JumpSlash, 0.1f);
 
-	//コリジョンの準備
-	m_attackCollision = NewGO<CollisionObject>(0);
-	m_attackCollision->CreateBox(enemy->GetPosition(), Quaternion::Identity, Vector3(50.0f, 50.0f, 50.0f));
-	m_attackCollision->SetName(ENEMY_ATTACK_COLLISION_NAME);
-	//横攻撃なので追加情報に横攻撃を設定
-	m_attackCollision->SetAdditionalInformation(ENEMY_SIDE_ATTACK_NAME);
-	m_attackCollision->SetIsEnableAutoDelete(false);
-
 	//コリジョンを生成するボーンのIDを取得
 	m_attackBoneID = enemy->GetModel()->FindBoneID(ENEMY_ATTACK_COLLISION_BONE_NAME);
-	//コリジョンのワールド行列を、取得したボーンのワールド行列に設定
-	m_attackCollision->SetWorldMatrix(enemy->GetModel()->GetBone(m_attackBoneID)->GetWorldMatrix());
 
 	m_hitFlag = false;
 
@@ -31,11 +22,16 @@ void EnemyStateSideSlash::Start(Enemy* enemy)
 
 void EnemyStateSideSlash::End(Enemy* enemy)
 {
-	//コリジョンを破棄
-	DeleteGO(m_attackCollision);
+	if (m_isAttackCollisionExistence)
+	{
+		//コリジョンを破棄
+		DeleteGO(m_attackCollision);
+
+		m_isAttackCollisionExistence = false;
+	}
 }
 
-void EnemyStateSideSlash::Animation(ModelRender& model, EnEnemyAnimationEvent animeEvent)
+void EnemyStateSideSlash::Animation(ModelRender& model, EnEnemyAnimationEvent& animeEvent)
 {
 	if (model.IsPlayingAnimation())
 	{
@@ -46,17 +42,39 @@ void EnemyStateSideSlash::Animation(ModelRender& model, EnEnemyAnimationEvent an
 		m_isPlayAnimation = false;
 	}
 
-	
+	switch (animeEvent)
+	{
+	case enEnemyAnimationEvent_AttackEnd:
+		animeEvent = enEnemyAnimationEvent_None;
+		//攻撃コリジョンを破棄
+		DeleteGO(m_attackCollision);
+
+		m_isAttackCollisionExistence = false;
+		break;
+
+	case enEnemyAnimationEvent_AttackStart:
+		animeEvent = enEnemyAnimationEvent_None;
+		//攻撃コリジョンを用意
+		m_attackCollision = NewGO<CollisionObject>(0, "enemyattack");
+		m_attackCollision->CreateBox(Vector3::Zero, Quaternion::Identity, ENEMY_ATTACK_COLLISION_SIZE);
+		m_attackCollision->SetName(ENEMY_ATTACK_COLLISION_NAME);
+		//縦攻撃なので追加情報に縦攻撃を設定
+		m_attackCollision->SetAdditionalInformation(ENEMY_SIDE_ATTACK_NAME);
+		m_attackCollision->SetIsEnableAutoDelete(false);
+
+		m_isAttackCollisionExistence = true;
+		break;
+
+	default:
+		animeEvent = enEnemyAnimationEvent_None;
+		break;
+	}
 }
 
 void EnemyStateSideSlash::Collision(const Vector3& pos, ModelRender& model, CharacterController& characon)
 {
-	//コリジョンのワールド行列を更新
-	m_attackCollision->SetWorldMatrix(model.GetBone(m_attackBoneID)->GetWorldMatrix());
-
-
 	//プレイヤーの攻撃コリジョン取得
-	const auto& AttackCollisions = g_collisionObjectManager->FindCollisionObjects("player_attack");
+	const auto& AttackCollisions = g_collisionObjectManager->FindCollisionObjects(PLAYER_ATTACK_COLLISION_NAME);
 
 	//被ダメージ判定
 	for (CollisionObject* collision : AttackCollisions)
@@ -67,6 +85,12 @@ void EnemyStateSideSlash::Collision(const Vector3& pos, ModelRender& model, Char
 
 			m_hitFlag = true;
 		}
+	}
+
+	if (m_isAttackCollisionExistence)
+	{
+		//コリジョンのワールド行列を更新
+		m_attackCollision->SetWorldMatrix(model.GetBone(m_attackBoneID)->GetWorldMatrix());
 	}
 }
